Reorganize project
- restructure project with distinct F53OSC directory separate from F53OSC Monitor targets
- add targets to manually build F53OSC.framework
- update .xcodeproj to Xcode 10.3
