+++ F53OSC Monitor/ActivityChartView.m
