+++ F53OSC Monitor/AppDelegate.h
@@ -3,11 +3,11 @@
 //  F53OSC Monitor
 //
 //  Created by Adam Bachman on 8/5/15.
-//  Copyright (c) 2015 Figure 53. All rights reserved.
+//  Copyright (c) 2015-2020 Figure 53. All rights reserved.
 //
 
 #import <Cocoa/Cocoa.h>
-#import "F53OSC.h"
+#import <F53OSC/F53OSC.h>
 #import "DemoServer.h"
 #import "ActivityChartView.h"
