+++ F53OSC Monitor/Images.xcassets/AppIcon.appiconset/Contents.json
