+++ F53OSC.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:F53OSC Monitor.xcodeproj">
+      location = "self:/Users/balord/Development/F53OSC/F53OSC.xcodeproj">
    </FileRef>
 </Workspace>
