+++ F53OSC.xcodeproj/xcshareddata/xcschemes/F53OSC Monitor.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "0640"
+   LastUpgradeVersion = "1030"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -17,39 +17,37 @@
                BlueprintIdentifier = "66EE17511B729EA0008B6743"
                BuildableName = "F53OSC Monitor.app"
                BlueprintName = "F53OSC Monitor"
-               ReferencedContainer = "container:F53OSC Monitor.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "NO"
-            buildForArchiving = "NO"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "66EE17641B729EA0008B6743"
-               BuildableName = "F53OSC MonitorTests.xctest"
-               BlueprintName = "F53OSC MonitorTests"
-               ReferencedContainer = "container:F53OSC Monitor.xcodeproj">
+               ReferencedContainer = "container:F53OSC.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
    </BuildAction>
    <TestAction
+      buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
-      shouldUseLaunchSchemeArgsEnv = "YES"
-      buildConfiguration = "Debug">
+      shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
          <TestableReference
-            skipped = "NO">
+            skipped = "NO"
+            testExecutionOrdering = "random">
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "66EE17641B729EA0008B6743"
                BuildableName = "F53OSC MonitorTests.xctest"
                BlueprintName = "F53OSC MonitorTests"
-               ReferencedContainer = "container:F53OSC Monitor.xcodeproj">
+               ReferencedContainer = "container:F53OSC.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
+         <TestableReference
+            skipped = "NO"
+            testExecutionOrdering = "random">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "3D1E0846242A7F5B00655E76"
+               BuildableName = "F53OSCTests.xctest"
+               BlueprintName = "F53OSCTests"
+               ReferencedContainer = "container:F53OSC.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -59,18 +57,21 @@
             BlueprintIdentifier = "66EE17511B729EA0008B6743"
             BuildableName = "F53OSC Monitor.app"
             BlueprintName = "F53OSC Monitor"
-            ReferencedContainer = "container:F53OSC Monitor.xcodeproj">
+            ReferencedContainer = "container:F53OSC.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
+      <AdditionalOptions>
+      </AdditionalOptions>
    </TestAction>
    <LaunchAction
+      buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       launchStyle = "0"
       useCustomWorkingDirectory = "NO"
-      buildConfiguration = "Debug"
       ignoresPersistentStateOnLaunch = "NO"
       debugDocumentVersioning = "YES"
+      debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
       <BuildableProductRunnable
          runnableDebuggingMode = "0">
@@ -79,17 +80,17 @@
             BlueprintIdentifier = "66EE17511B729EA0008B6743"
             BuildableName = "F53OSC Monitor.app"
             BlueprintName = "F53OSC Monitor"
-            ReferencedContainer = "container:F53OSC Monitor.xcodeproj">
+            ReferencedContainer = "container:F53OSC.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
    <ProfileAction
+      buildConfiguration = "Release"
       shouldUseLaunchSchemeArgsEnv = "YES"
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
-      buildConfiguration = "Release"
       debugDocumentVersioning = "YES">
       <BuildableProductRunnable
          runnableDebuggingMode = "0">
@@ -98,7 +99,7 @@
             BlueprintIdentifier = "66EE17511B729EA0008B6743"
             BuildableName = "F53OSC Monitor.app"
             BlueprintName = "F53OSC Monitor"
-            ReferencedContainer = "container:F53OSC Monitor.xcodeproj">
+            ReferencedContainer = "container:F53OSC.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
