+++ F53OSC/Tests/F53OSCTests/F53OSCMessageTests.m
@@ -13,7 +13,7 @@
 #import <Foundation/Foundation.h>
 #import <XCTest/XCTest.h>
 
-#import "F53OSC.h"
+#import <F53OSC/F53OSC.h>
 
 
 NS_ASSUME_NONNULL_BEGIN
