+++ README.md
@@ -2,7 +2,7 @@
 
 Hey neat, it's a nice open source OSC library for Objective-C.
 
-From your friends at [Figure 53](http://figure53.com).
+From your friends at [Figure 53](https://figure53.com).
 
 For convenience, we've included a few public domain source files from [CocoaAsyncSocket](https://github.com/robbiehanson/CocoaAsyncSocket).  But appropriate thanks, kudos, and curiosity about that code should be directed to [the source](https://github.com/robbiehanson/CocoaAsyncSocket).
 
@@ -10,7 +10,7 @@ For convenience, we've included a few public domain source files from [CocoaAsyn
 
 - F53OSC must be compiled with ARC.
 - You must link against `Security.framework` and `CFNetwork.framework`.
-- F53OSC requires Xcode 6.3 or later and a deployment target of macOS 10.9+ or iOS 8.4+.
+- F53OSC requires Xcode 10.3 or later and a deployment target of macOS 10.9+, iOS 8.4+, or tvOS 9.0+.
 
 You can also use CocoaPods to include F53OSC into your project:
 
@@ -20,4 +20,4 @@ pod 'F53OSC', :git => 'https://github.com/Figure53/F53OSC.git'
 
 ## Demo
 
-Inluded is a small demo app "F53OSC Monitor", which logs OSC messages sent to it via port 9999 and displays some basic stats about incoming traffic.
+Included is a small demo app "F53OSC Monitor", which logs OSC messages sent to it via port 9999 and displays some basic stats about incoming traffic.
